+++ challenge/urls.py
@@ -1,8 +1,10 @@
 from django.contrib import admin
 from django.urls import path, include
-from rest_framework import routers
-from register.views import ClientViewSet, CompanyViewSet
-from rest_framework.authtoken import views
+from django.urls.conf import re_path
+from rest_framework.authentication import BaseAuthentication
+from rest_framework.routers import DefaultRouter
+from accounts.views import ClientViewSet, CompanyViewSet, UserViewSet, ManagerViewSet, LoginViewSet
+
 
 from rest_framework import permissions
 from drf_yasg.views import get_schema_view
@@ -12,9 +14,12 @@ from django.conf import settings
 from django.conf.urls.static import static
 
 
-router = routers.DefaultRouter()
-router.register(r'Client', ClientViewSet, basename= 'Client')
-router.register(r'Company', CompanyViewSet, basename='Company')
+router = DefaultRouter()
+router.register(r'client', ClientViewSet, basename='client')
+router.register(r'company', CompanyViewSet, basename='company')
+router.register(r'user', UserViewSet, basename='user')
+router.register(r'manager', ManagerViewSet, basename='manager')
+router.register(r'login', LoginViewSet, basename='login')
 
 
 schema_view = get_schema_view(
@@ -32,12 +37,8 @@ schema_view = get_schema_view(
 
 
 urlpatterns = [
-    path('admin/', admin.site.urls),
-    path('document/meusarquivos',admin.site.urls),
-    path('api-auth/', include('rest_framework.urls', namespace='rest_framework')),
-    path('', include(router.urls)),
-    path('swagger/', schema_view.with_ui('swagger', cache_timeout=0), name='schema-swagger-ui'),
-
+    re_path(r'^v1/admin/', admin.site.urls),
+    re_path(r'^v1/api-auth/', include('rest_framework.urls', namespace='v1')),
+    re_path(r'^v1/', include(router.urls)),
+    re_path(r'^v1/swagger/', schema_view.with_ui('swagger', cache_timeout=0), name = 'schema-swagger-ui'),
 ] + static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
-
-STATIC_ROOT = "/var/www/example.com/static/"
